Revert rename to `entities`
