+++ src/classes/osu/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from classes.osu.Beatmap import Beatmap
+from classes.osu.enums import *
+from classes.osu.Mods import Mods
+from classes.osu.Performance import BeatmapPerformance
+from classes.osu.Performance import Performance
+from classes.osu.Score import BeatmapScore
+from classes.osu.Score import RecentScore
+from classes.osu.Score import Score
+from classes.osu.User import User
