+++ src/classes/osu/enums/BeatmapIcon.py
