+++ src/classes/osu/enums/BeatmapStatus.py
