+++ src/classes/osu/enums/Mod.py
