+++ src/classes/osu/enums/Mode.py
