+++ src/classes/osu/enums/Server.py
