+++ src/classes/osu/enums/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from classes.osu.enums.BeatmapIcon import BeatmapIcon
+from classes.osu.enums.BeatmapStatus import BeatmapStatus
+from classes.osu.enums.Mod import Mod
+from classes.osu.enums.Mode import Mode
+from classes.osu.enums.Rank import Rank
+from classes.osu.enums.Server import Server
