+++ src/commons/osu/osuhelpers.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-import entities.osu as osu
+import classes.osu as osu
 from commons.regex import beatmap_id_rx
 from commons.regex import beatmap_link_rx
