--- src/entities/osu/__init__.py
@@ -1,11 +1,0 @@
-from __future__ import annotations
-
-from entities.osu.Beatmap import Beatmap
-from entities.osu.enums import *
-from entities.osu.Mods import Mods
-from entities.osu.Performance import BeatmapPerformance
-from entities.osu.Performance import Performance
-from entities.osu.Score import BeatmapScore
-from entities.osu.Score import RecentScore
-from entities.osu.Score import Score
-from entities.osu.User import User
