--- src/entities/osu/enums/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-from entities.osu.enums.BeatmapIcon import BeatmapIcon
-from entities.osu.enums.BeatmapStatus import BeatmapStatus
-from entities.osu.enums.Mod import Mod
-from entities.osu.enums.Mode import Mode
-from entities.osu.enums.Rank import Rank
-from entities.osu.enums.Server import Server
