+++ src/run.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import logging
 
 import sentry_sdk
-from entities.bot import Sunny
+from classes.bot import Sunny
 
 logger = logging.getLogger()
 bot = Sunny()
